Move uploads api to User controller/service Closes: #12 Closes: #175
