--- WebClient/src/services/pagination.ts
@@ -1,8 +1,0 @@
-// No real better place to put this... We should generate declarations for all API models.
-export interface IPaginationMetadata {
-    count: number;
-
-    previous: string;
-
-    next: string;
-}
