+++ Website/Src/Controllers/UserController.cs
@@ -10,6 +10,8 @@ namespace Website.Controllers
     using System.Text.RegularExpressions;
     using System.Threading.Tasks;
     using ClickerHeroesTrackerWebsite.Models;
+    using ClickerHeroesTrackerWebsite.Models.Api;
+    using ClickerHeroesTrackerWebsite.Models.Api.Uploads;
     using ClickerHeroesTrackerWebsite.Models.Api.Users;
     using ClickerHeroesTrackerWebsite.Models.Game;
     using ClickerHeroesTrackerWebsite.Models.Settings;
@@ -76,6 +78,133 @@ namespace Website.Controllers
             }
 
             return this.BadRequest(this.ModelState);
+        }
+
+        [Route("{userName}/uploads")]
+        [HttpGet]
+        [AllowAnonymous]
+        public async Task<IActionResult> List(
+            string userName,
+            int page = ParameterConstants.List.Page.Default,
+            int count = ParameterConstants.List.Count.Default)
+        {
+            if (string.IsNullOrEmpty(userName))
+            {
+                return this.BadRequest();
+            }
+
+            var user = await this.userManager.FindByNameAsync(userName);
+            if (user == null)
+            {
+                return this.NotFound();
+            }
+
+            var userId = await this.userManager.GetUserIdAsync(user);
+            if (string.IsNullOrEmpty(userId))
+            {
+                return this.NotFound();
+            }
+
+            var currentUserId = this.userManager.GetUserId(this.User);
+            if (!userId.Equals(currentUserId, StringComparison.OrdinalIgnoreCase)
+                && !this.User.IsInRole("Admin"))
+            {
+                return this.Forbid();
+            }
+
+            // Validate parameters
+            if (page < ParameterConstants.List.Page.Min)
+            {
+                return this.BadRequest();
+            }
+
+            if (count < ParameterConstants.List.Count.Min
+                || count > ParameterConstants.List.Count.Max)
+            {
+                return this.BadRequest();
+            }
+
+            const string GetUploadsCommandText = @"
+                SELECT Id, UploadTime
+                FROM Uploads
+                WHERE UserId = @UserId
+                ORDER BY UploadTime DESC
+                OFFSET @Offset ROWS
+                FETCH NEXT @Count ROWS ONLY;";
+            var getUploadsCommandparameters = new Dictionary<string, object>
+            {
+                { "@UserId", userId },
+                { "@Offset", (page - 1) * count },
+                { "@Count", count },
+            };
+
+            var uploads = new List<Upload>(count);
+            using (var command = this.databaseCommandFactory.Create(GetUploadsCommandText, getUploadsCommandparameters))
+            using (var reader = command.ExecuteReader())
+            {
+                while (reader.Read())
+                {
+                    uploads.Add(new Upload
+                    {
+                        Id = Convert.ToInt32(reader["Id"]),
+
+                        // The DateTime is a datetime2 which has no timezone so comes out as DateTimeKind.Unknown. Se need to specify the kind so it gets serialized correctly.
+                        TimeSubmitted = DateTime.SpecifyKind(Convert.ToDateTime(reader["UploadTime"]), DateTimeKind.Utc),
+                    });
+                }
+            }
+
+            const string GetUploadCountCommandText = @"
+                SELECT COUNT(*) AS TotalUploads
+                FROM Uploads
+                WHERE UserId = @UserId";
+            var getUploadCountparameters = new Dictionary<string, object>
+            {
+                { "@UserId", userId },
+            };
+
+            var pagination = new PaginationMetadata();
+            using (var command = this.databaseCommandFactory.Create(GetUploadCountCommandText, getUploadCountparameters))
+            using (var reader = command.ExecuteReader())
+            {
+                if (!reader.Read())
+                {
+                    return null;
+                }
+
+                pagination.Count = Convert.ToInt32(reader["TotalUploads"]);
+
+                var currentPath = this.Request.Path;
+                if (page > 1)
+                {
+                    pagination.Previous = string.Format(
+                        "{0}?{1}={2}&{3}={4}",
+                        currentPath,
+                        nameof(page),
+                        page - 1,
+                        nameof(count),
+                        count);
+                }
+
+                if (page <= Math.Ceiling((float)pagination.Count / count))
+                {
+                    pagination.Next = string.Format(
+                        "{0}?{1}={2}&{3}={4}",
+                        currentPath,
+                        nameof(page),
+                        page + 1,
+                        nameof(count),
+                        count);
+                }
+            }
+
+            var model = new UploadSummaryListResponse()
+            {
+                Uploads = uploads,
+                Pagination = pagination,
+            };
+
+            return this.Ok(model);
         }
 
         [Route("{userName}/progress")]
@@ -192,12 +321,12 @@ namespace Website.Controllers
                 { "@UserId", userId },
             };
             const string GetUserFollowsCommandText = @"
-	            SELECT UserName
-	            FROM UserFollows
-	            INNER JOIN AspNetUsers
-	            ON UserFollows.FollowUserId = AspNetUsers.Id
-	            WHERE UserId = @UserId
-	            ORDER BY UserName ASC";
+                SELECT UserName
+                FROM UserFollows
+                INNER JOIN AspNetUsers
+                ON UserFollows.FollowUserId = AspNetUsers.Id
+                WHERE UserId = @UserId
+                ORDER BY UserName ASC";
             using (var command = this.databaseCommandFactory.Create(
                 GetUserFollowsCommandText,
                 parameters))
@@ -523,6 +652,28 @@ namespace Website.Controllers
             }
 
             return this.BadRequest(this.ModelState);
+        }
+
+        internal static class ParameterConstants
+        {
+            internal static class List
+            {
+                internal static class Page
+                {
+                    internal const int Min = 1;
+
+                    internal const int Default = 1;
+                }
+
+                internal static class Count
+                {
+                    internal const int Min = 1;
+
+                    internal const int Max = 100;
+
+                    internal const int Default = 10;
+                }
+            }
         }
     }
 }
